Move engine tests to its on directory
